refactor: merge hooks and stores
