--- src/data/hooks/index.ts
@@ -1,7 +1,0 @@
-export {
-    useSettings,
-    useUpdateSettings,
-    useCommands,
-    useValidCommands,
-    useFindCommand
-} from "@/data/stores/useSettingStore";
